Create RadiusLifecycle object and move some logic into it
The main thing here is having an object to pass around instead of a bare
callback that keeps Chewie far too coupled to the RADIUS implementation.
A bit more work to be done in the test_radius suite, although this shows
we probably need to think about what the Radius -> RadiusLifecycle
relationship looks like.

There's a little more to trim out of Chewie.receive_radius_messages()
too, and that should make Chewie's responsibility a lot more obvious and
easier to test.